fix: Replace secretHandshake with a more deterministic implementation
Instead of "secretHandshake", NetworkDiscovery now uses a deterministic "uniqueAppidentifier" which is generated by the  GetUniqueAppIdentifier() virtual method. This ID is generated based on project name and version, which makes it a more reliable implementation. Old "secretHandshake" used a randomly generated value, which could be easily lost if the NetworkDiscovery component was reset. 
The new GetUniqueAppIdentifier() provides a deterministic identifier for the whole project, and can be overridden in child classes shall the need arise.