fix: Fix infinite loop when stopping discovery from response callback
Setting UdpClient references to null was causing an Editor freeze if it happened in ProcessServerResponse() callback. However, there is no need to set those to null explicitly, as UdpClient is already getting disposed of properly when client.Close() is called.
This fixes freezes in an expected use case where the discovery is stopped once the server is found (from inside the ProcessServerResponse() callback).